requirements.txt
absl-py==0.2.0
alabaster==0.7.10
astor==0.6.2
Babel==2.5.3
beautifulsoup4==4.6.0
bleach==1.5.0
boto==2.48.0
boto3==1.7.14
botocore==1.10.14
bz2file==0.98
certifi==2018.4.16
chardet==3.0.4
click==6.7
cssselect==1.0.3
cycler==0.10.0
decorator==4.3.0
docutils==0.14
feedfinder2==0.0.4
feedparser==5.2.1
Flask==1.0.2
gast==0.2.0
gensim==0.13.4.1
grpcio==1.11.0
html5lib==0.9999999
idna==2.6
imagesize==1.0.0
itsdangerous==0.24
jieba3k==0.35.1
Jinja2==2.10
jmespath==0.9.3
llvmlite==0.23.0
lxml==4.2.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.0.0
networkx==1.11
newspaper3k==0.2.6
nltk==3.2.2
numba==0.30.1
numpy==1.14.3
oauthlib==2.0.7
pandas==0.19.2
pexpect==4.2.1
Pillow==5.1.0
protobuf==3.5.2.post1
ptyprocess==0.5.2
pycorenlp==0.3.0
Pygments==2.2.0
pyparsing==2.2.0
PySocks==1.6.8
python-dateutil==2.7.2
pytz==2018.4
PyYAML==3.12
regex==2016.12.27
requests==2.18.4
requests-file==1.4.3
requests-oauthlib==0.8.0
s3transfer==0.1.13
scikit-learn==0.18.1
scipy==0.18.1
six==1.11.0
smart-open==1.5.7
snowballstemmer==1.2.1
Sphinx==1.5.1
tensorboard==1.8.0
tensorflow==1.8.0
termcolor==1.1.0
tldextract==2.2.0
tqdm==4.11.1
tweepy==3.6.0
urllib3==1.22
Werkzeug==0.14.1
